backend/requirements.txt
fastapi
uvicorn[standard]
pillow
pvl
requests
tqdm
